backend/requirements.txt
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
pydantic>=2.6.0
pydantic-settings>=2.1.0
python-multipart>=0.0.9
httpx>=0.26.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
structlog>=24.1.0
tenacity>=8.2.3
sqlalchemy>=2.0.25
aiosqlite>=0.19.0
alembic>=1.13.1
